Extract the storage sheet into a store and dialog
Move the storage sheet out of MainWindow. A StorageStore owns the non-UI
operations — detecting which selected releases have a transition in flight,
running a transition across a selection (including the unmanage folder pick and
the in-flight tracking), and intersecting the actions a selection allows; the
action intersection delegates to a unit-tested OrderedIntersection. The
StorageDialog renders the release table, outbox, and download queue, and now
registers its own live-refresh handlers on the projection registry while open
instead of MainWindow holding refresh-callback fields.
