+++ bae-windows/MainWindow.xaml.cs
@@ -53,24 +53,13 @@ public sealed partial class MainWindow : Window
     private readonly ReleaseActionDialogs _releaseActions;
     private readonly AlbumDetailDialog _albumDetail;
     private readonly QueueDialog _queueDialog;
-
-    // Releases whose unmanage is running right now. Unmanage is a blocking
-    // foreground transfer (unlike pin, which enqueues, or upload, which lives in
-    // the outbox), so it has no queue snapshot to read — we track it here while
-    // RunStorageActionForReleases awaits, letting the storage row offer to cancel
-    // it. UI-thread only (added/removed around the await, read when building the
-    // menu).
-    private readonly HashSet<string> _unmanagingReleases = new();
+    private readonly StorageStore _storage;
+    private readonly StorageDialog _storageDialog;
 
     // Drives the system media transport controls (hardware media keys + the
     // Windows media flyout) from playback events. One instance for the window's
     // lifetime; library switches deactivate it rather than recreating it.
     private readonly MediaControlService _mediaControls;
-
-    // Reloads the storage dialog's outbox panel and storage rows; set while that
-    // dialog is open so outbox invalidations refresh them live, null when closed.
-    private Action? _refreshOutbox;
-    private Action? _refreshDownloads;
 
     // Re-reads generated bridge settings into the open settings dialog's labels; set while that
     // dialog is open so config invalidations refresh it live, null when closed.
@@ -80,11 +69,6 @@ public sealed partial class MainWindow : Window
     // the manual check/download/apply from settings. Inert on a dev run or a
     // loose-zip copy (IsAvailable is false).
     private readonly UpdateService _updateService = new();
-
-    // Reloads the storage dialog's release rows; set while that dialog is open so
-    // album/release invalidations refresh each row's state badge and actions, not
-    // just the album grid.
-    private Action? _refreshStorageRows;
 
     // x:Bind Albums/Composers in the shell resolve here; the collections live in
     // the browser store (constructed before InitializeComponent so these are
@@ -176,6 +160,16 @@ public sealed partial class MainWindow : Window
             _releaseActions);
         _queueDialog = new QueueDialog(_session, _playback, () => Content.XamlRoot);
 
+        // The storage sheet and its non-UI operations. The dialog registers its
+        // live-refresh handlers on the projection registry while open.
+        _storage = new StorageStore(_session);
+        _storageDialog = new StorageDialog(
+            _session,
+            () => Content.XamlRoot,
+            () => WinRT.Interop.WindowNative.GetWindowHandle(this),
+            _storage,
+            _projections);
+
         // The composer/search panes and the library-lifecycle dialogs. The window
         // stays the navigation shell (open/close/switch); these render and drive
         // their own screens, calling back for the operations it owns.
@@ -246,12 +240,8 @@ public sealed partial class MainWindow : Window
     {
         _projections.Register(typeof(BridgeInvalidation.AlbumList), ReloadBrowserFromInvalidation);
         _projections.Register(typeof(BridgeInvalidation.ComposerList), ReloadBrowserFromInvalidation);
-        _projections.Register(typeof(BridgeInvalidation.Album), () => _refreshStorageRows?.Invoke());
-        _projections.Register(typeof(BridgeInvalidation.Release), () => _refreshStorageRows?.Invoke());
         _projections.Register(typeof(BridgeInvalidation.Config), () => _refreshSettings?.Invoke());
         _projections.Register(typeof(BridgeInvalidation.SyncStatus), _sync.Refresh);
-        _projections.Register(typeof(BridgeInvalidation.Outbox), () => _refreshOutbox?.Invoke());
-        _projections.Register(typeof(BridgeInvalidation.DownloadQueue), () => _refreshDownloads?.Invoke());
         _projections.Register(typeof(BridgeInvalidation.ImportCandidateList), _import.RefreshCandidates);
         _projections.Register(typeof(BridgeInvalidation.ImportCandidate), _import.RefreshCandidates);
         _projections.Register(typeof(BridgeInvalidation.WatchedFolders), _import.RefreshCandidates);
@@ -818,839 +808,7 @@ public sealed partial class MainWindow : Window
 
     private async void OnStorageClick(object sender, RoutedEventArgs e)
     {
-        if (CurrentHandleOrNull() == null)
-        {
-            return;
-        }
-
-        var listPanel = new StackPanel { Spacing = 4, MinWidth = 460 };
-        var storageStatus = new TextBlock
-        {
-            Foreground = new SolidColorBrush(Microsoft.UI.Colors.Salmon),
-            TextWrapping = TextWrapping.Wrap,
-            Visibility = Visibility.Collapsed,
-        };
-
-        // The releases whose rows are selected. Right-clicking applies the
-        // chosen action to the whole selection (or to just the right-tapped row
-        // when it isn't part of it). Releases that vanish on reload (e.g. an
-        // local release moved out of the library) drop out below.
-        var selected = new HashSet<string>();
-        // The current rows, kept so the right-tap menu can resolve a release's
-        // allowed actions (for the multi-select intersection) by id.
-        var rowsById = new Dictionary<string, BridgeStorageRow>();
-
-        // Each row shows its summary; a left-click toggles its selection and a
-        // right-click opens a menu of the transitions the core says it allows
-        // (carried on the row, gated on cloud-home + pending uploads), plus
-        // cancel for any queued uploads. The same actions run on every selected
-        // release.
-        async System.Threading.Tasks.Task LoadStorageRows()
-        {
-            var (current, result) = await RunForCurrentHandle(NativeBae.StorageRows);
-            if (!current)
-            {
-                return;
-            }
-            if (result.Error is not null)
-            {
-                storageStatus.Text = result.Error;
-                storageStatus.Visibility = Visibility.Visible;
-                return;
-            }
-            if (result.Rows is null)
-            {
-                storageStatus.Text = Loc.Chrome("storage.load_failed");
-                storageStatus.Visibility = Visibility.Visible;
-                return;
-            }
-
-            storageStatus.Visibility = Visibility.Collapsed;
-            rowsById.Clear();
-            foreach (var row in result.Rows)
-            {
-                rowsById[row.Release.Id] = row;
-            }
-            // Drop selections for releases no longer present after a transition.
-            selected.IntersectWith(rowsById.Keys);
-
-            listPanel.Children.Clear();
-            foreach (var row in result.Rows)
-            {
-                var text = new TextBlock
-                {
-                    Text = StorageRowSummary(row),
-                    VerticalAlignment = VerticalAlignment.Center,
-                    TextWrapping = TextWrapping.Wrap,
-                };
-                var releaseId = row.Release.Id;
-                var rowBorder = new Border
-                {
-                    Child = text,
-                    // The release id rides on Tag so RefreshRowHighlights can
-                    // recolor each row from the current selection.
-                    Tag = releaseId,
-                    Padding = new Thickness(6, 4, 6, 4),
-                    CornerRadius = new CornerRadius(4),
-                    Background = RowBackground(selected.Contains(releaseId)),
-                };
-
-                rowBorder.Tapped += (_, _) =>
-                {
-                    if (!selected.Add(releaseId))
-                    {
-                        selected.Remove(releaseId);
-                    }
-                    rowBorder.Background = RowBackground(selected.Contains(releaseId));
-                };
-                rowBorder.RightTapped += async (_, args) =>
-                {
-                    // The args are only valid synchronously; capture the tap
-                    // position before any await.
-                    var position = args.GetPosition(rowBorder);
-
-                    // Act on the selection when this row is part of it, else on
-                    // just this row (and select it, matching the macOS menu).
-                    if (!selected.Contains(releaseId))
-                    {
-                        selected.Clear();
-                        selected.Add(releaseId);
-                        RefreshRowHighlights();
-                    }
-
-                    var menu = await BuildStorageRowMenu(
-                        selected.ToList(), rowsById, storageStatus, LoadStorageRows);
-                    // Nothing to offer (e.g. no cloud home, or uploads in flight)
-                    // — skip the empty popup.
-                    if (menu.Items.Count > 0)
-                    {
-                        menu.ShowAt(rowBorder, new FlyoutShowOptions { Position = position });
-                    }
-                };
-
-                listPanel.Children.Add(rowBorder);
-            }
-        }
-
-        // Repaint every row's background from the current selection. The storage
-        // list is a flat StackPanel of Borders tagged with their release id.
-        void RefreshRowHighlights()
-        {
-            foreach (var child in listPanel.Children)
-            {
-                if (child is Border border && border.Tag is string id)
-                {
-                    border.Background = RowBackground(selected.Contains(id));
-                }
-            }
-        }
-
-        await LoadStorageRows();
-
-        // Cloud outbox: the upload/delete queue with a summary band, a Retry-now
-        // button, and per-item Cancel. Hidden (empty panel) when nothing is queued.
-        // Reloaded after retry/cancel so the panel reflects the new queue state.
-        var downloadsPanel = new StackPanel { Spacing = 4 };
-        async System.Threading.Tasks.Task LoadDownloads()
-        {
-            downloadsPanel.Children.Clear();
-            var (current, result) = await RunForCurrentHandle(NativeBae.DownloadSnapshot);
-            if (!current)
-            {
-                return;
-            }
-            if (result.Error is not null)
-            {
-                storageStatus.Text = result.Error;
-                storageStatus.Visibility = Visibility.Visible;
-                return;
-            }
-            var snapshot = result.Snapshot;
-            if (snapshot is null)
-            {
-                storageStatus.Text = Loc.Chrome("storage.read_failed");
-                storageStatus.Visibility = Visibility.Visible;
-                return;
-            }
-
-            // Hidden when the pin queue is idle, like the outbox panel.
-            if (snapshot.Downloads.Length == 0)
-            {
-                return;
-            }
-
-            string StateLabel(BridgeDownloadOp op) => op.State switch
-            {
-                BridgeDownloadState.Active => Loc.Chrome("download.state.downloading"),
-                BridgeDownloadState.Failed => Loc.Chrome("download.state.failed"),
-                _ => Loc.Chrome("download.state.queued"),
-            };
-
-            string DownloadDetail(BridgeDownloadOp op)
-            {
-                static string DisplayBytes(ulong bytes) =>
-                    Loc.Bytes(checked((long)bytes));
-
-                var parts = new List<string>
-                {
-                    Loc.Chrome("storage.files", "count", op.FileCount),
-                    Loc.Bytes(op.TotalSize),
-                    StateLabel(op),
-                };
-                if (DownloadProgress(op.State) is { } progress)
-                {
-                    parts.Add(
-                        Loc.Core(
-                            "core.download.bytes_progress",
-                            new Dictionary<string, object?>
-                            {
-                                ["done"] = DisplayBytes(progress.BytesDone),
-                                ["total"] = DisplayBytes(progress.BytesTotal),
-                            }));
-                }
-                return string.Join(" · ", parts);
-            }
-
-            // Header: a label (or "paused"), Retry (only with failures), and a
-            // pause/resume toggle — mirroring the outbox panel's band.
-            var band = new StackPanel { Orientation = Orientation.Horizontal, Spacing = 8 };
-            band.Children.Add(new TextBlock
-            {
-                Text = snapshot.Paused ? Loc.Chrome("download.paused") : Loc.Chrome("download.title"),
-                VerticalAlignment = VerticalAlignment.Center,
-            });
-            var retry = new Button
-            {
-                Content = Loc.Chrome("outbox.retry_now"),
-                IsEnabled = snapshot.Total.Failed > 0,
-            };
-            retry.Click += async (_, _) =>
-            {
-                retry.IsEnabled = false;
-                await RunForCurrentHandle(NativeBae.RetryDownloads);
-                await LoadDownloads();
-            };
-            band.Children.Add(retry);
-            var paused = snapshot.Paused;
-            var pause = new Button { Content = paused ? Loc.Chrome("outbox.resume") : Loc.Chrome("outbox.pause") };
-            pause.Click += async (_, _) =>
-            {
-                pause.IsEnabled = false;
-                await RunForCurrentHandle(
-                    handle => NativeBae.SetDownloadsPaused(handle, !paused));
-                await LoadDownloads();
-            };
-            band.Children.Add(pause);
-            downloadsPanel.Children.Add(band);
-
-            // One row per release: title, "N files · size · state", and a cancel.
-            foreach (var op in snapshot.Downloads)
-            {
-                var itemGrid = new Grid { ColumnSpacing = 8 };
-                itemGrid.ColumnDefinitions.Add(new ColumnDefinition { Width = new GridLength(1, GridUnitType.Star) });
-                itemGrid.ColumnDefinitions.Add(new ColumnDefinition { Width = GridLength.Auto });
-
-                var labelColumn = new StackPanel { Spacing = 2, VerticalAlignment = VerticalAlignment.Center };
-                labelColumn.Children.Add(new TextBlock { Text = op.Title, TextWrapping = TextWrapping.Wrap });
-                labelColumn.Children.Add(new TextBlock
-                {
-                    Text = DownloadDetail(op),
-                    Foreground = new SolidColorBrush(Microsoft.UI.Colors.Gray),
-                });
-                if (DownloadProgress(op.State) is { } progress)
-                {
-                    labelColumn.Children.Add(new ProgressBar
-                    {
-                        Minimum = 0,
-                        Maximum = 1,
-                        Value = progress.Fraction,
-                        Height = 4,
-                    });
-                }
-                Grid.SetColumn(labelColumn, 0);
-                itemGrid.Children.Add(labelColumn);
-
-                var releaseId = op.ReleaseId;
-                var cancel = new Button { Content = Loc.Chrome("action.cancel") };
-                cancel.Click += async (_, _) =>
-                {
-                    storageStatus.Visibility = Visibility.Collapsed;
-                    cancel.IsEnabled = false;
-                    var (cancelCurrent, error) = await RunForCurrentHandle(
-                        handle => NativeBae.CancelReleaseTransition(handle, releaseId));
-                    if (!cancelCurrent)
-                    {
-                        return;
-                    }
-                    if (error is not null)
-                    {
-                        storageStatus.Text = error;
-                        storageStatus.Visibility = Visibility.Visible;
-                        cancel.IsEnabled = true;
-                        return;
-                    }
-
-                    await LoadDownloads();
-                };
-                Grid.SetColumn(cancel, 1);
-                itemGrid.Children.Add(cancel);
-                downloadsPanel.Children.Add(itemGrid);
-            }
-        }
-
-        var outboxPanel = new StackPanel { Spacing = 4 };
-        async System.Threading.Tasks.Task LoadOutbox()
-        {
-            outboxPanel.Children.Clear();
-            var (current, result) = await RunForCurrentHandle(NativeBae.OutboxSnapshot);
-            if (!current)
-            {
-                return;
-            }
-            if (result.Error is not null)
-            {
-                storageStatus.Text = result.Error;
-                storageStatus.Visibility = Visibility.Visible;
-                return;
-            }
-            var snapshot = result.Snapshot;
-            if (snapshot is null)
-            {
-                storageStatus.Text = Loc.Chrome("outbox.load_failed");
-                storageStatus.Visibility = Visibility.Visible;
-                return;
-            }
-
-            if (snapshot.UploadGroups.Length == 0 && snapshot.Deletes.Length == 0)
-            {
-                return;
-            }
-
-            // With work queued at least one count is non-zero, so compose the
-            // localized queue summary from the generated snapshot counts.
-            var band = new StackPanel { Orientation = Orientation.Horizontal, Spacing = 8 };
-            band.Children.Add(new TextBlock
-            {
-                Text = OutboxSummary(snapshot),
-                VerticalAlignment = VerticalAlignment.Center,
-            });
-            var retry = new Button { Content = Loc.Chrome("outbox.retry_now") };
-            retry.Click += async (_, _) =>
-            {
-                storageStatus.Visibility = Visibility.Collapsed;
-                retry.IsEnabled = false;
-                var (retryCurrent, error) = await RunForCurrentHandle(NativeBae.RetryOutbox);
-                if (!retryCurrent)
-                {
-                    return;
-                }
-                if (error is not null)
-                {
-                    storageStatus.Text = error;
-                    storageStatus.Visibility = Visibility.Visible;
-                    retry.IsEnabled = true;
-                    return;
-                }
-
-                await LoadOutbox();
-            };
-            band.Children.Add(retry);
-            // Pause/resume the upload pipeline. Paused leaves items queued but stops
-            // the sync cycle from draining them.
-            var paused = snapshot.Paused;
-            var pause = new Button { Content = paused ? Loc.Chrome("outbox.resume") : Loc.Chrome("outbox.pause") };
-            pause.Click += async (_, _) =>
-            {
-                pause.IsEnabled = false;
-                await RunForCurrentHandle(handle => NativeBae.SetSyncPaused(handle, !paused));
-                await LoadOutbox();
-            };
-            band.Children.Add(pause);
-            outboxPanel.Children.Add(band);
-
-            // Master progress strip: a byte-progress bar (dimmed while paused) and
-            // locale-formatted byte / throughput / ETA labels.
-            if (snapshot.Total.BytesTotal > 0)
-            {
-                outboxPanel.Children.Add(new ProgressBar
-                {
-                    Minimum = 0,
-                    Maximum = checked((long)snapshot.Total.BytesTotal),
-                    Value = checked((long)snapshot.Total.BytesDone),
-                    Opacity = paused ? 0.4 : 1.0,
-                });
-                var detail = new List<string>();
-                var bytesLabel = OutboxBytesLabel(snapshot);
-                if (!string.IsNullOrEmpty(bytesLabel))
-                {
-                    detail.Add(bytesLabel);
-                }
-                var throughputLabel = OutboxThroughputLabel(snapshot);
-                if (!string.IsNullOrEmpty(throughputLabel))
-                {
-                    detail.Add(throughputLabel);
-                }
-                var etaLabel = OutboxEtaLabel(snapshot);
-                if (!string.IsNullOrEmpty(etaLabel))
-                {
-                    detail.Add(etaLabel);
-                }
-                if (detail.Count > 0)
-                {
-                    outboxPanel.Children.Add(new TextBlock
-                    {
-                        Text = string.Join(" · ", detail),
-                        FontSize = 12,
-                        Foreground = new SolidColorBrush(Microsoft.UI.Colors.Gray),
-                    });
-                }
-            }
-
-            // A queue row: a label (with an optional progress bar), an optional
-            // trailing button, and an optional right-click menu.
-            void AddOutboxRow(string label, ProgressBar? progress, Button? trailing, MenuFlyout? contextMenu)
-            {
-                var itemGrid = new Grid { ColumnSpacing = 8 };
-                itemGrid.ColumnDefinitions.Add(new ColumnDefinition { Width = new GridLength(1, GridUnitType.Star) });
-                itemGrid.ColumnDefinitions.Add(new ColumnDefinition { Width = GridLength.Auto });
-
-                var labelColumn = new StackPanel { Spacing = 2, VerticalAlignment = VerticalAlignment.Center };
-                labelColumn.Children.Add(new TextBlock { Text = label, TextWrapping = TextWrapping.Wrap });
-                if (progress is not null)
-                {
-                    labelColumn.Children.Add(progress);
-                }
-                Grid.SetColumn(labelColumn, 0);
-                itemGrid.Children.Add(labelColumn);
-
-                if (trailing is not null)
-                {
-                    Grid.SetColumn(trailing, 1);
-                    itemGrid.Children.Add(trailing);
-                }
-                if (contextMenu is not null)
-                {
-                    itemGrid.ContextFlyout = contextMenu;
-                }
-                outboxPanel.Children.Add(itemGrid);
-            }
-
-            // Runs `action` off-thread, surfaces any error to the status line, and
-            // reloads the panel on success — shared by the row button and menu.
-            async System.Threading.Tasks.Task RunCancel(Func<AppHandle, string?> action)
-            {
-                storageStatus.Visibility = Visibility.Collapsed;
-                var (current, error) = await RunForCurrentHandle(action);
-                if (!current)
-                {
-                    return;
-                }
-                if (error is not null)
-                {
-                    storageStatus.Text = error;
-                    storageStatus.Visibility = Visibility.Visible;
-                    return;
-                }
-
-                await LoadOutbox();
-            }
-
-            // A right-click "Cancel" menu, matching the storage table's per-release
-            // cancel. Used for the upload release rows.
-            MenuFlyout CancelFlyout(Func<AppHandle, string?> action)
-            {
-                var menu = new MenuFlyout();
-                var item = new MenuFlyoutItem { Text = Loc.Chrome("action.cancel") };
-                item.Click += async (_, _) => await RunCancel(action);
-                menu.Items.Add(item);
-                return menu;
-            }
-
-            // Uploads: one row per release (matching the storage table) — title,
-            // aggregate progress, and a right-click "Cancel" that stops the
-            // release's transition. The orphaned-files bucket (no release id) has
-            // no release to cancel.
-            foreach (var group in snapshot.UploadGroups)
-            {
-                ProgressBar? progress = group.Progress is { Active: > 0, BytesTotal: > 0 }
-                    ? new ProgressBar
-                    {
-                        Minimum = 0,
-                        Maximum = checked((long)group.Progress.BytesTotal),
-                        Value = checked((long)group.Progress.BytesDone),
-                    }
-                    : null;
-                MenuFlyout? menu = group.ReleaseId is string releaseId
-                    ? CancelFlyout(handle => NativeBae.CancelReleaseTransition(handle, releaseId))
-                    : null;
-                AddOutboxRow(group.DisplayTitle, progress, trailing: null, contextMenu: menu);
-            }
-            // A pending delete is genuinely a single-file operation, so it keeps
-            // its own per-file cancel button.
-            foreach (var delete in snapshot.Deletes)
-            {
-                var cancel = new Button { Content = Loc.Chrome("outbox.cancel_item") };
-                var id = delete.Id;
-                cancel.Click += async (_, _) => await RunCancel(
-                    handle => NativeBae.CancelOutboxItem(handle, id));
-                AddOutboxRow(DeleteLabel(delete), null, trailing: cancel, contextMenu: null);
-            }
-        }
-
-        await LoadDownloads();
-        await LoadOutbox();
-
-        var content = new StackPanel { Spacing = 8 };
-        content.Children.Add(storageStatus);
-        content.Children.Add(downloadsPanel);
-        content.Children.Add(outboxPanel);
-        content.Children.Add(listPanel);
-
-        var dialog = new ContentDialog
-        {
-            Title = Loc.Chrome("storage.title"),
-            Content = new ScrollViewer { Content = content, MaxHeight = 480 },
-            CloseButtonText = Loc.Chrome("action.close"),
-            XamlRoot = Content.XamlRoot,
-        };
-        // Refresh both the outbox panel and the storage rows live while the dialog
-        // is open as uploads/deletes progress. Stops once the dialog closes.
-        _refreshOutbox = () =>
-        {
-            _ = LoadOutbox();
-            _ = LoadStorageRows();
-        };
-        // Refresh the Downloads pane live as pins progress and the storage rows
-        // with them (a row's badge/state changes as a pin completes).
-        _refreshDownloads = () =>
-        {
-            _ = LoadDownloads();
-            _ = LoadStorageRows();
-        };
-        // An album/release invalidation that isn't an outbox change can still
-        // alter a release's storage state — refresh the rows for it too.
-        _refreshStorageRows = () => _ = LoadStorageRows();
-        try
-        {
-            await dialog.ShowAsync();
-        }
-        finally
-        {
-            _refreshOutbox = null;
-            _refreshDownloads = null;
-            _refreshStorageRows = null;
-        }
-    }
-
-    // Selected-row highlight: a faint accent tint, or transparent when not
-    // selected. Static so LoadStorageRows and RefreshRowHighlights agree.
-    private static Brush RowBackground(bool isSelected) =>
-        isSelected
-            ? new SolidColorBrush(Microsoft.UI.Colors.SteelBlue) { Opacity = 0.25 }
-            : new SolidColorBrush(Microsoft.UI.Colors.Transparent);
-
-    // User-facing label for a storage transition, matching the macOS
-    // "Storage…" sheet / context menu wording.
-    private static string StorageRowSummary(BridgeStorageRow row)
-    {
-        var format = string.IsNullOrEmpty(row.Release.Format) ? string.Empty : $" · {row.Release.Format}";
-        var files = Loc.Chrome("storage.files", "count", row.Release.FileCount);
-        return $"{row.Album.Title} — {row.Album.ArtistNames}{format} · {files} · {Loc.Bytes(row.Release.TotalSize)} · {StorageStateLabel(row.Release.StorageState)}{PinIndicator(row.Release)}";
-    }
-
-    private static string StorageStateLabel(BridgeReleaseStorageState state) => state switch
-    {
-        BridgeReleaseStorageState.Remote => Loc.Chrome("storage.state.managed"),
-        BridgeReleaseStorageState.Local => Loc.Chrome("storage.state.unmanaged"),
-        _ => throw new ArgumentOutOfRangeException(nameof(state), state, "Unknown storage state"),
-    };
-
-    private static string PinIndicator(BridgeReleaseSummary release) =>
-        release.Pinned ? $" · {Loc.Chrome("storage.pinned")}" : string.Empty;
-
-    private static BridgeDownloadTransferProgress? DownloadProgress(BridgeDownloadState state) =>
-        state is BridgeDownloadState.Active active ? active.Progress : null;
-
-    private static string OutboxSummary(BridgeOutboxSnapshot snapshot)
-    {
-        var parts = new List<string>();
-        if (snapshot.Total.Active > 0) parts.Add(Loc.Core("core.queue.uploading", "count", snapshot.Total.Active));
-        if (snapshot.Total.Failed > 0) parts.Add(Loc.Core("core.queue.failed", "count", snapshot.Total.Failed));
-        if (snapshot.Total.Queued > 0) parts.Add(Loc.Core("core.queue.queued", "count", snapshot.Total.Queued));
-        if (snapshot.Deletes.Length > 0)
-            parts.Add(Loc.Core("core.outbox.pending_deletes", "count", snapshot.Deletes.Length));
-        return string.Join(" · ", parts);
-    }
-
-    private static string OutboxThroughputLabel(BridgeOutboxSnapshot snapshot) =>
-        snapshot.ThroughputBps > 0
-            ? Loc.Core("core.outbox.throughput", "rate", Loc.Bytes(checked((long)snapshot.ThroughputBps)))
-            : string.Empty;
-
-    private static string OutboxEtaLabel(BridgeOutboxSnapshot snapshot) =>
-        snapshot.EtaSeconds is { } seconds
-            ? Loc.Core("core.outbox.eta", "duration", Loc.Duration(checked(checked((long)seconds) * 1000)))
-            : string.Empty;
-
-    private static string OutboxBytesLabel(BridgeOutboxSnapshot snapshot)
-    {
-        if (snapshot.Total.BytesTotal == 0) return string.Empty;
-        return Loc.Core(
-            "core.outbox.bytes_progress",
-            new Dictionary<string, object?>
-            {
-                ["done"] = Loc.Bytes(checked((long)snapshot.Total.BytesDone)),
-                ["total"] = Loc.Bytes(checked((long)snapshot.Total.BytesTotal)),
-            });
-    }
-
-    private static string DeleteLabel(BridgeDeleteOp delete) =>
-        $"{delete.CloudKey} — {Loc.Chrome("outbox.delete.kind")}";
-
-    private static string StorageActionLabel(BridgeReleaseStorageAction action) => action switch
-    {
-        BridgeReleaseStorageAction.MakeRemote => Loc.Chrome("storage.action.manage"),
-        BridgeReleaseStorageAction.MakeLocal => Loc.Chrome("storage.action.unmanage"),
-        BridgeReleaseStorageAction.Pin => Loc.Chrome("storage.action.pin"),
-        BridgeReleaseStorageAction.Unpin => Loc.Chrome("storage.action.unpin"),
-        _ => throw new ArgumentOutOfRangeException(nameof(action), action, "Unknown storage action"),
-    };
-
-    // The transitions every release in the selection allows, intersected so the
-    // menu only offers actions applicable to all. Order follows the first
-    // release's action list (the core's order). The caller suppresses actions
-    // when any targeted release has a transition in flight, matching the macOS
-    // "Storage…" sheet.
-    private static List<BridgeReleaseStorageAction> IntersectedStorageActions(
-        List<string> releaseIds, Dictionary<string, BridgeStorageRow> rowsById)
-    {
-        var perRelease = releaseIds
-            .Select(id => rowsById.TryGetValue(id, out var row)
-                ? new HashSet<BridgeReleaseStorageAction>(row.Release.StorageActions)
-                : new HashSet<BridgeReleaseStorageAction>())
-            .ToList();
-        if (perRelease.Count == 0)
-        {
-            return new List<BridgeReleaseStorageAction>();
-        }
-
-        var common = perRelease[0];
-        foreach (var set in perRelease.Skip(1))
-        {
-            common.IntersectWith(set);
-        }
-        // Preserve the core's action order from the first release's row.
-        var order = rowsById.TryGetValue(releaseIds[0], out var firstRow)
-            ? firstRow.Release.StorageActions
-            : [];
-        return order.Where(common.Contains).ToList();
-    }
-
-    // Build the right-tap menu for the targeted releases: the intersected
-    // storage transitions plus a cancel for any of their queued uploads. Each
-    // item runs the action on every targeted release, then reloads the rows.
-    private async System.Threading.Tasks.Task<MenuFlyout> BuildStorageRowMenu(
-        List<string> releaseIds,
-        Dictionary<string, BridgeStorageRow> rowsById,
-        TextBlock storageStatus,
-        Func<System.Threading.Tasks.Task> reload)
-    {
-        var menu = new MenuFlyout();
-
-        // A release with a transition in flight offers only "Cancel" — the
-        // storage actions (pin/unmanage/…) would race it. Each transition
-        // surfaces differently: an upload sits in the outbox snapshot, a pin in
-        // the download queue snapshot, and an unmanage (a blocking foreground
-        // transfer with no queue) is tracked locally while it runs. Core
-        // dispatches to whichever is running.
-        var transitioning = new HashSet<string>(
-            await UploadingReleases(releaseIds, storageStatus));
-        transitioning.UnionWith(await DownloadingReleases(releaseIds));
-        transitioning.UnionWith(releaseIds.Where(_unmanagingReleases.Contains));
-        if (transitioning.Count > 0)
-        {
-            var cancel = new MenuFlyoutItem { Text = Loc.Chrome("action.cancel") };
-            cancel.Click += async (_, _) =>
-            {
-                foreach (var releaseId in transitioning)
-                {
-                    var (current, error) = await RunForCurrentHandle(
-                        handle => NativeBae.CancelReleaseTransition(handle, releaseId));
-                    if (!current)
-                    {
-                        return;
-                    }
-                    if (error is not null)
-                    {
-                        storageStatus.Text = error;
-                        storageStatus.Visibility = Visibility.Visible;
-                        return;
-                    }
-                }
-
-                await reload();
-            };
-            menu.Items.Add(cancel);
-            return menu;
-        }
-
-        foreach (var action in IntersectedStorageActions(releaseIds, rowsById))
-        {
-            var act = action;
-            var item = new MenuFlyoutItem { Text = StorageActionLabel(act) };
-            item.Click += async (_, _) =>
-            {
-                var error = await RunStorageActionForReleases(act, releaseIds);
-                if (error is not null)
-                {
-                    storageStatus.Text = error;
-                    storageStatus.Visibility = Visibility.Visible;
-                }
-                else
-                {
-                    await reload();
-                }
-            };
-            menu.Items.Add(item);
-        }
-
-        return menu;
-    }
-
-    // Of the given releases, those with uploads queued or in flight. Core omits
-    // idle releases from the per-release map, so presence there is the signal.
-    private async System.Threading.Tasks.Task<List<string>> UploadingReleases(
-        List<string> releaseIds,
-        TextBlock storageStatus)
-    {
-        var (current, result) = await RunForCurrentHandle(NativeBae.OutboxSnapshot);
-        if (!current)
-        {
-            return new List<string>();
-        }
-        if (result.Error is not null)
-        {
-            storageStatus.Text = result.Error;
-            storageStatus.Visibility = Visibility.Visible;
-            return new List<string>();
-        }
-        var snapshot = result.Snapshot;
-        if (snapshot is null)
-        {
-            // Couldn't read the outbox; surface it like the panel load does
-            // rather than silently dropping the cancel action.
-            storageStatus.Text = Loc.Chrome("outbox.read_failed");
-            storageStatus.Visibility = Visibility.Visible;
-            return new List<string>();
-        }
-
-        return releaseIds.Where(snapshot.PerRelease.ContainsKey).ToList();
-    }
-
-    // Of the given releases, those queued or downloading in the pin queue.
-    private async System.Threading.Tasks.Task<List<string>> DownloadingReleases(
-        List<string> releaseIds)
-    {
-        var (current, result) = await RunForCurrentHandle(NativeBae.DownloadSnapshot);
-        if (!current)
-        {
-            return new List<string>();
-        }
-        if (result.Error is not null)
-        {
-            BaeDiagnostics.Logger.Warning(
-                $"couldn't read the download snapshot; pin-cancel unavailable: {result.Error}");
-            return new List<string>();
-        }
-        var snapshot = result.Snapshot;
-        if (snapshot is null)
-        {
-            // The pin queue is in-memory and read is infallible bar a dropped
-            // handle, so this is an app-state fault, not a user-facing read
-            // error — log it and offer no pin-cancel rather than a toast.
-            BaeDiagnostics.Logger.Warning(
-                "couldn't read the download snapshot; pin-cancel unavailable");
-            return new List<string>();
-        }
-
-        var pinning = snapshot.Downloads.Select(op => op.ReleaseId).ToHashSet();
-        return releaseIds.Where(pinning.Contains).ToList();
-    }
-
-    // Run a storage transition on every release in the selection off the UI
-    // thread. "unmanage" asks once for a destination folder, then moves each
-    // release into it. Returns null on success (or a cancelled picker), else the
-    // first error message.
-    private async System.Threading.Tasks.Task<string?> RunStorageActionForReleases(
-        BridgeReleaseStorageAction action, List<string> releaseIds)
-    {
-        if (action == BridgeReleaseStorageAction.MakeLocal)
-        {
-            var picker = new global::Windows.Storage.Pickers.FolderPicker();
-            picker.FileTypeFilter.Add("*");
-            WinRT.Interop.InitializeWithWindow.Initialize(
-                picker, WinRT.Interop.WindowNative.GetWindowHandle(this));
-            var folder = await picker.PickSingleFolderAsync();
-            if (folder is null)
-            {
-                return null;
-            }
-
-            var path = folder.Path;
-            // Mark these releases as unmanaging so a right-click can cancel them
-            // while the blocking transfer runs; cleared when it returns.
-            foreach (var releaseId in releaseIds)
-            {
-                _unmanagingReleases.Add(releaseId);
-            }
-            try
-            {
-                var (storageActionCurrent, error) = await RunForCurrentHandle(handle =>
-                {
-                    foreach (var releaseId in releaseIds)
-                    {
-                        var error = NativeBae.MakeReleaseLocal(handle, releaseId, path);
-                        if (error is not null)
-                        {
-                            return error;
-                        }
-                    }
-
-                    return (string?)null;
-                });
-                return storageActionCurrent ? error : null;
-            }
-            finally
-            {
-                foreach (var releaseId in releaseIds)
-                {
-                    _unmanagingReleases.Remove(releaseId);
-                }
-            }
-        }
-
-        var (current, actionError) = await RunForCurrentHandle(handle =>
-        {
-            foreach (var releaseId in releaseIds)
-            {
-                var error = action switch
-                {
-                    BridgeReleaseStorageAction.Pin => NativeBae.PinRelease(handle, releaseId),
-                    BridgeReleaseStorageAction.Unpin => NativeBae.UnpinRelease(handle, releaseId),
-                    BridgeReleaseStorageAction.MakeRemote => NativeBae.MakeReleaseRemote(handle, releaseId, pin: false),
-                    BridgeReleaseStorageAction.MakeLocal => throw new InvalidOperationException(
-                        "make-local storage actions must choose a destination before running"),
-                    _ => throw new ArgumentOutOfRangeException(nameof(action), action, "Unknown storage action"),
-                };
-                if (error is not null)
-                {
-                    return error;
-                }
-            }
-
-            return (string?)null;
-        });
-        return current ? actionError : null;
+        await _storageDialog.Show();
     }
 
     private async void OnSettingsClick(object sender, RoutedEventArgs e)
